% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimargins.R
\name{mimargins}
\alias{mimargins}
\title{Calculation of Average Marginal Effects for Multiple Imputed Data in R}
\usage{
mimargins(
  df,
  formula,
  imp_var = ".imp",
  type = "lm",
  ame = T,
  variables = NULL,
  at = NULL,
  linkfunction = NULL
)
}
\arguments{
\item{df}{The data frame which should be used for the regressions. Should be in long format and preferably not a mids-object. Whether mids-objects work has not yet been tested, however.}

\item{formula}{The Regression formula. Should preferably be a formula object. Must correspond to regression type specified in 'type'.}

\item{imp_var}{The variable in the data frame indicating the imputation (m). Defaults to ".imp".}

\item{type}{Specifies which type of regression to run. Currently supports 'lm' (Default), 'lmer' and 'glm'. Has to correspond to the specified formula.}

\item{ame}{Whether average marginal effects (AMEs) shall be calculated. Defaults to TRUE. If set to FALSE, a regression with the formula above is run for each imputation and pooled using Rubin's Rules. If set to true, AMEs are calculated for each imputation and then pooled using Rubin's rules.}

\item{variables}{Whether AMEs shall be calculated for a specific variable only. Currently only works for specifying one variable. Only has an effect if ame = T. Defaults to NULL. Then, AMEs are calculated.}

\item{at}{Whether AMEs shall be calculated at specified levels of another variable. Currently only works for specifying one variable. The number of levels may vary. Input has to take the form list(VARIABLENAME = c(level1, level2, level3 ...)). Only has an effect if ame = T. Defaults to NULL. Then, AMEs are averaged over all levels of all other variables.}

\item{linkfunction}{If type is set to 'glm', a linkfunction needs to be specified. Has no effect if type is set to other values.}
}
\description{
Calculation of Average Marginal Effects for Multiple Imputed Data in R
}
