% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipredictions.R
\name{mipredictions}
\alias{mipredictions}
\title{Predicted Values Calculation}
\usage{
mipredictions(
  df,
  formula,
  imp_var = ".imp",
  id_var,
  type = "lm",
  at = NULL,
  linkfunction = NULL
)
}
\arguments{
\item{df}{The data frame which should be used for the regressions. Should be in long format and preferably not a mids-object. Whether mids-objects work has not yet been tested, however.}

\item{formula}{The Regression formula. Should preferably be a formula object. Must correspond to regression type specified in 'type'.}

\item{imp_var}{The variable in the data frame indicating the imputation (m). Defaults to ".imp".}

\item{type}{Specifies which type of regression to run. Currently supports 'lm' (Default), 'lmer' and 'glm'. Has to correspond to the specified formula.}

\item{at}{Whether AMEs shall be calculated at specified levels of another variable. Currently only works for specifying one variable. The number of levels may vary. Input has to take the form list(VARIABLENAME = c(level1, level2, level3 ...)). Defaults to NULL. Then, Predicted Values are averaged over all levels of all other variables.}

\item{linkfunction}{If type is set to 'glm', a linkfunction needs to be specified. Has no effect if type is set to other values.}
}
\value{
Returns Datafram of Predicted Values and SEs.
}
\description{
Predicted Values Calculation
}
